+++ src/storage/buffer_manager.cpp
@@ -33,7 +33,7 @@ bool Frame::acquire(bool block) {
 
 BufferManager::BufferManager(uint64_t maxSize)
     : logger{spdlog::stdout_logger_mt("buffer_manager")},
-      bufferCache{maxSize / PAGE_SIZE}, clockHand{0}, numFrames{maxSize / PAGE_SIZE} {
+      bufferCache{maxSize / PAGE_SIZE}, clockHand{0}, numFrames{(uint32_t)(maxSize / PAGE_SIZE)} {
     logger->info("Initializing Buffer Manager.");
     logger->info("BufferPool Size {}B, #4KB-pages {}.", maxSize, maxSize / PAGE_SIZE);
     logger->info("Done.");
